+++ src/dominio/PotenciacaoParser.kt
@@ -11,8 +11,8 @@ class PotenciacaoParser : Parser() {
         val matchResult = regex.find(expressao) ?: return ""
         val groups = matchResult.groups
 
-        val primeiroNumero: Double = trataNumeroDeGruoDeCaptura(groups, 1).toDouble()
-        val segundoNumero: Double = trataNumeroDeGruoDeCaptura(groups, 4).toDouble()
+        val primeiroNumero: Double = trataNumeroDeGruoDeCaptura(groups, 1)
+        val segundoNumero: Double = trataNumeroDeGruoDeCaptura(groups, 4)
 
         val resultadoParcial: String = if (groups[1]?.value == "(-") {
             (-1 * primeiroNumero.pow(segundoNumero)).toString()
